requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
decorator==5.1.1
Django==4.2.2
django-appconf==1.0.5
django-bootstrap-form==3.4
django-crispy-forms==2.0
django-multiselectfield==0.1.12
django-redis==5.3.0
django-widget-tweaks==1.4.12
glob2==0.7
idna==3.4
Pillow==9.5.0
python-memcached==1.59
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.4
validators==0.20.0
